Extend PhpDocPropertySorterFixer to support @property-read and @property-write
